feat: enhance category management with image upload and validation
- Updated category model to include structured image data with URL and fileId.
- Modified createCategory and updateCategory functions to handle image uploads via ImageKit.
- Implemented image deletion when updating or deleting categories.
- Enhanced category routes to support image upload using multer middleware.
- Updated validation schemas to accommodate optional title and description fields.

+++ server/src/services/category.service.js
@@ -1,9 +1,27 @@
 import Category from "../models/category.model.js";
 import ApiError from "../utils/ApiError.js";
+import {
+  uploadImageToImageKit,
+  deleteImageFromImageKit,
+} from "../services/imagekit.service.js";
 
 export const createCategory = async (categoryDetails) => {
   const { title, description, image } = categoryDetails;
-  const category = await Category.create({ title, description, image });
+
+  if (!image) {
+    throw ApiError.badRequestError("Image is required");
+  }
+  const imageKitResponse = await uploadImageToImageKit(
+    image,
+    title,
+    "foodie/categories"
+  );
+
+  const category = await Category.create({
+    title,
+    description,
+    image: imageKitResponse,
+  });
   return category;
 };
 
@@ -18,13 +36,43 @@ export const getCategoryById = async (id) => {
 };
 
 export const updateCategory = async (id, categoryDetails) => {
-  const category = await Category.findByIdAndUpdate(id, categoryDetails, {
-    new: true,
-  });
-  return category;
+  const category = await Category.findById(id);
+  if (!category) {
+    throw ApiError.notFoundError("Category");
+  }
+  if (categoryDetails.image) {
+    const imageKitResponse = await uploadImageToImageKit(
+      categoryDetails.image,
+      category.title,
+      "foodie/categories"
+    );
+    categoryDetails.image = imageKitResponse;
+
+    // delete old image
+    if (category.image?.fileId) {
+      await deleteImageFromImageKit(category.image.fileId);
+    }
+  }
+  const updatedCategory = await Category.findByIdAndUpdate(
+    id,
+    categoryDetails,
+    {
+      new: true,
+    }
+  );
+  return updatedCategory;
 };
 
 export const deleteCategory = async (id) => {
-  const category = await Category.findByIdAndDelete(id);
-  return category;
+  const category = await Category.findById(id)
+  if (!category) {
+    throw ApiError.notFoundError("Category");
+  }
+
+  // delete image
+  if (category.image?.fileId) {
+    await deleteImageFromImageKit(category.image.fileId);
+  }
+
+  return await Category.findByIdAndDelete(id);
 };
